+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Add support for selling multiples in shops. Kind of hacky right now, like "carrot_item@64: 100"
  - Add spell-overrides region flag, for region-specific spell tweaks (e.g. broom max height)
  - Add mob protection to Towny integration
+ - Added "console" parameter and @p, @pn etc replacement to Recall commands
 
 ## 5.3.1